refactor package structure for contract tests
